refactor: files-and-videos folder (#672)
