+++ src/files-and-videos/files-page/FilesPage.test.jsx
@@ -7,6 +7,8 @@ import {
   within,
 } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import ReactDOM from 'react-dom';
+import { saveAs } from 'file-saver';
 
 import { initializeMockApp } from '@edx/frontend-platform';
 import MockAdapter from 'axios-mock-adapter';
@@ -17,39 +19,37 @@ import { IntlProvider } from '@edx/frontend-platform/i18n';
 import initializeStore from '../../store';
 import { executeThunk } from '../../utils';
 import { RequestStatus } from '../../data/constants';
-import Videos from './Videos';
+import FilesPage from './FilesPage';
 import {
-  generateFetchVideosApiResponse,
+  generateFetchAssetApiResponse,
   generateEmptyApiResponse,
-  generateNewVideoApiResponse,
-  generateAddVideoApiResponse,
+  generateNewAssetApiResponse,
   getStatusValue,
   courseId,
   initialState,
 } from './factories/mockApiResponses';
 
 import {
-  fetchVideos,
-  addVideoFile,
-  deleteVideoFile,
+  fetchAssets,
+  addAssetFile,
+  deleteAssetFile,
+  updateAssetLock,
   getUsagePaths,
-  addVideoThumbnail,
-  fetchVideoDownload,
 } from './data/thunks';
-import { getVideosUrl, getCoursVideosApiUrl, getApiBaseUrl } from './data/api';
-import videoMessages from './messages';
-import messages from '../messages';
+import { getAssetsUrl } from './data/api';
+import messages from '../generic/messages';
 
 let axiosMock;
 let store;
 let file;
+ReactDOM.createPortal = jest.fn(node => node);
 jest.mock('file-saver');
 
 const renderComponent = () => {
   render(
     <IntlProvider locale="en">
       <AppProvider store={store}>
-        <Videos courseId={courseId} />
+        <FilesPage courseId={courseId} />
       </AppProvider>
     </IntlProvider>,
   );
@@ -58,15 +58,15 @@ const renderComponent = () => {
 const mockStore = async (
   status,
 ) => {
-  const fetchVideosUrl = getVideosUrl(courseId);
-  axiosMock.onGet(fetchVideosUrl).reply(getStatusValue(status), generateFetchVideosApiResponse());
-  await executeThunk(fetchVideos(courseId), store.dispatch);
+  const fetchAssetsUrl = `${getAssetsUrl(courseId)}?page_size=50`;
+  axiosMock.onGet(fetchAssetsUrl).reply(getStatusValue(status), generateFetchAssetApiResponse());
+  await executeThunk(fetchAssets(courseId), store.dispatch);
 };
 
 const emptyMockStore = async (status) => {
-  const fetchVideosUrl = getVideosUrl(courseId);
-  axiosMock.onGet(fetchVideosUrl).reply(getStatusValue(status), generateEmptyApiResponse());
-  await executeThunk(fetchVideos(courseId), store.dispatch);
+  const fetchAssetsUrl = getAssetsUrl(courseId);
+  axiosMock.onGet(fetchAssetsUrl).reply(getStatusValue(status), generateEmptyApiResponse());
+  await executeThunk(fetchAssets(courseId), store.dispatch);
 };
 
 describe('FilesAndUploads', () => {
@@ -82,14 +82,14 @@ describe('FilesAndUploads', () => {
       });
       store = initializeStore({
         ...initialState,
-        videos: {
-          ...initialState.videos,
-          videoIds: [],
+        assets: {
+          ...initialState.assets,
+          assetIds: [],
         },
         models: {},
       });
       axiosMock = new MockAdapter(getAuthenticatedHttpClient());
-      file = new File(['(⌐□_□)'], 'download.mp4', { type: 'video/mp4' });
+      file = new File(['(⌐□_□)'], 'download.png', { type: 'image/png' });
     });
 
     it('should return placeholder component', async () => {
@@ -98,16 +98,10 @@ describe('FilesAndUploads', () => {
       expect(screen.getByTestId('under-construction-placeholder')).toBeVisible();
     });
 
-    it('should not render transcript settings button', async () => {
+    it('should have Files title', async () => {
       renderComponent();
       await emptyMockStore(RequestStatus.SUCCESSFUL);
-      expect(screen.queryByText(videoMessages.transcriptSettingsButtonLabel.defaultMessage));
-    });
-
-    it('should have Video uploads title', async () => {
-      renderComponent();
-      await emptyMockStore(RequestStatus.SUCCESSFUL);
-      expect(screen.getByText(videoMessages.heading.defaultMessage)).toBeVisible();
+      expect(screen.getByText('Files')).toBeVisible();
     });
 
     it('should render dropzone', async () => {
@@ -123,20 +117,14 @@ describe('FilesAndUploads', () => {
       await emptyMockStore(RequestStatus.SUCCESSFUL);
       const dropzone = screen.getByTestId('files-dropzone');
       await act(async () => {
-        const mockResponseData = { status: '200', ok: true, blob: () => 'Data' };
-        const mockFetchResponse = Promise.resolve(mockResponseData);
-        global.fetch = jest.fn().mockImplementation(() => mockFetchResponse);
-
-        axiosMock.onPost(getCoursVideosApiUrl(courseId)).reply(204, generateNewVideoApiResponse());
-        axiosMock.onGet(getCoursVideosApiUrl(courseId)).reply(200, generateAddVideoApiResponse());
-
+        axiosMock.onPost(getAssetsUrl(courseId)).reply(204, generateNewAssetApiResponse());
         Object.defineProperty(dropzone, 'files', {
           value: [file],
         });
         fireEvent.drop(dropzone);
-        await executeThunk(addVideoFile(courseId, file), store.dispatch);
+        await executeThunk(addAssetFile(courseId, file, 0), store.dispatch);
       });
-      const addStatus = store.getState().videos.addingStatus;
+      const addStatus = store.getState().assets.addingStatus;
       expect(addStatus).toEqual(RequestStatus.SUCCESSFUL);
 
       expect(screen.queryByTestId('files-dropzone')).toBeNull();
@@ -145,7 +133,7 @@ describe('FilesAndUploads', () => {
     });
   });
 
-  describe('valid videos', () => {
+  describe('valid assets', () => {
     beforeEach(async () => {
       initializeMockApp({
         authenticatedUser: {
@@ -160,20 +148,11 @@ describe('FilesAndUploads', () => {
       file = new File(['(⌐□_□)'], 'download.png', { type: 'image/png' });
     });
 
+    afterEach(() => {
+      saveAs.mockClear();
+    });
+
     describe('table view', () => {
-      it('should render transcript settings button', async () => {
-        renderComponent();
-        await mockStore(RequestStatus.SUCCESSFUL);
-        const transcriptSettingsButton = screen.getByText(videoMessages.transcriptSettingsButtonLabel.defaultMessage);
-        expect(transcriptSettingsButton).toBeVisible();
-
-        await act(async () => {
-          fireEvent.click(transcriptSettingsButton);
-        });
-
-        expect(screen.getByLabelText('close settings')).toBeVisible();
-      });
-
       it('should render table with gallery card', async () => {
         renderComponent();
         await mockStore(RequestStatus.SUCCESSFUL);
@@ -199,39 +178,19 @@ describe('FilesAndUploads', () => {
 
         expect(screen.getByRole('table')).toBeVisible();
       });
-
-      it('should update video thumbnail', async () => {
-        renderComponent();
-        await mockStore(RequestStatus.SUCCESSFUL);
-        axiosMock.onPost(`${getApiBaseUrl()}/video_images/${courseId}/mOckID1`).reply(200, { image_url: 'url' });
-        const addThumbnailButton = screen.getByTestId('video-thumbnail-mOckID1');
-        const thumbnail = new File(['test'], 'sOMEUrl.jpg', { type: 'image/jpg' });
-        await act(async () => {
-          fireEvent.click(addThumbnailButton);
-          await executeThunk(addVideoThumbnail({ file: thumbnail, videoId: 'mOckID1', courseId }), store.dispatch);
-        });
-        const updateStatus = store.getState().videos.updatingStatus;
-        expect(updateStatus).toEqual(RequestStatus.SUCCESSFUL);
-      });
     });
 
     describe('table actions', () => {
       it('should upload a single file', async () => {
         renderComponent();
         await mockStore(RequestStatus.SUCCESSFUL);
-        const mockResponseData = { status: '200', ok: true, blob: () => 'Data' };
-        const mockFetchResponse = Promise.resolve(mockResponseData);
-        global.fetch = jest.fn().mockImplementation(() => mockFetchResponse);
-
-        axiosMock.onPost(getCoursVideosApiUrl(courseId)).reply(204, generateNewVideoApiResponse());
-        axiosMock.onGet(getCoursVideosApiUrl(courseId)).reply(200, generateAddVideoApiResponse());
-
-        const addFilesButton = screen.getAllByLabelText('file-input')[3];
+        axiosMock.onPost(getAssetsUrl(courseId)).reply(200, generateNewAssetApiResponse());
+        const addFilesButton = screen.getByLabelText('file-input');
         await act(async () => {
           userEvent.upload(addFilesButton, file);
-          await executeThunk(addVideoFile(courseId, file), store.dispatch);
+          await executeThunk(addAssetFile(courseId, file, 1), store.dispatch);
         });
-        const addStatus = store.getState().videos.addingStatus;
+        const addStatus = store.getState().assets.addingStatus;
         expect(addStatus).toEqual(RequestStatus.SUCCESSFUL);
       });
 
@@ -263,7 +222,7 @@ describe('FilesAndUploads', () => {
         const deleteButton = screen.getByText(messages.deleteTitle.defaultMessage).closest('a');
         expect(deleteButton).not.toHaveClass('disabled');
 
-        axiosMock.onDelete(`${getCoursVideosApiUrl(courseId)}/mOckID1`).reply(204);
+        axiosMock.onDelete(`${getAssetsUrl(courseId)}mOckID1`).reply(204);
 
         fireEvent.click(deleteButton);
         expect(screen.getByText(messages.deleteConfirmationTitle.defaultMessage)).toBeVisible();
@@ -282,8 +241,8 @@ describe('FilesAndUploads', () => {
 
         expect(screen.queryByText(messages.deleteConfirmationTitle.defaultMessage)).toBeNull();
 
-        // Check if the video is deleted in the store and UI
-        const deleteStatus = store.getState().videos.deletingStatus;
+        // Check if the asset is deleted in the store and UI
+        const deleteStatus = store.getState().assets.deletingStatus;
         expect(deleteStatus).toEqual(RequestStatus.SUCCESSFUL);
         expect(screen.queryByTestId('grid-card-mOckID1')).toBeNull();
       });
@@ -302,12 +261,8 @@ describe('FilesAndUploads', () => {
         const downloadButton = screen.getByText(messages.downloadTitle.defaultMessage).closest('a');
         expect(downloadButton).not.toHaveClass('disabled');
 
-        await act(async () => {
-          fireEvent.click(downloadButton);
-        });
-
-        const updateStatus = store.getState().videos.updatingStatus;
-        expect(updateStatus).toEqual(RequestStatus.SUCCESSFUL);
+        fireEvent.click(downloadButton);
+        expect(saveAs).toHaveBeenCalled();
       });
 
       it('download button should be enabled and download multiple selected files', async () => {
@@ -322,18 +277,14 @@ describe('FilesAndUploads', () => {
         await waitFor(() => {
           fireEvent.click(actionsButton);
         });
-        axiosMock.onGet(`${getVideosUrl(courseId)}/mOckID1`).reply(200, { download_link: 'http://download.org' });
-        axiosMock.onGet(`${getVideosUrl(courseId)}/mOckID5`).reply(200, { download_link: 'http://download.org' });
-
+        const mockResponseData = { ok: true, blob: () => 'Data' };
+        const mockFetchResponse = Promise.resolve(mockResponseData);
         const downloadButton = screen.getByText(messages.downloadTitle.defaultMessage).closest('a');
         expect(downloadButton).not.toHaveClass('disabled');
 
-        await act(async () => {
-          fireEvent.click(downloadButton);
-        });
-
-        const updateStatus = store.getState().videos.updatingStatus;
-        expect(updateStatus).toEqual(RequestStatus.SUCCESSFUL);
+        global.fetch = jest.fn().mockImplementation(() => mockFetchResponse);
+        fireEvent.click(downloadButton);
+        expect(fetch).toHaveBeenCalledTimes(2);
       });
 
       it('sort button should be enabled and sort files by name', async () => {
@@ -372,73 +323,105 @@ describe('FilesAndUploads', () => {
     });
 
     describe('card menu actions', () => {
-      it('should open video info', async () => {
+      it('should open asset info', async () => {
         renderComponent();
         await mockStore(RequestStatus.SUCCESSFUL);
         expect(screen.getByTestId('grid-card-mOckID1')).toBeVisible();
 
-        const videoMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
-        expect(videoMenuButton).toBeVisible();
+        const assetMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
+        expect(assetMenuButton).toBeVisible();
 
-        axiosMock.onGet(`${getVideosUrl(courseId)}/mOckID1/usage`)
-          .reply(201, { usageLocations: ['subsection - unit / block'] });
+        axiosMock.onGet(`${getAssetsUrl(courseId)}mOckID1/usage`).reply(201, { usageLocations: ['subsection - unit / block'] });
         await waitFor(() => {
-          fireEvent.click(within(videoMenuButton).getByLabelText('file-menu-toggle'));
+          fireEvent.click(within(assetMenuButton).getByLabelText('file-menu-toggle'));
           fireEvent.click(screen.getByText('Info'));
+          executeThunk(getUsagePaths({
+            courseId,
+            asset: { id: 'mOckID1', displayName: 'mOckID1' },
+            setSelectedRows: jest.fn(),
+          }), store.dispatch);
+          expect(screen.getAllByLabelText('mOckID1')[0]).toBeVisible();
         });
 
-        expect(screen.getByText(messages.infoTitle.defaultMessage)).toBeVisible();
-
-        const { usageStatus } = store.getState().videos;
-
+        const { usageStatus } = store.getState().assets;
         expect(usageStatus).toEqual(RequestStatus.SUCCESSFUL);
-
         expect(screen.getByText('subsection - unit / block')).toBeVisible();
       });
 
-      it('should open video info modal and show info tab', async () => {
+      it('should open asset info and handle lock checkbox', async () => {
         renderComponent();
         await mockStore(RequestStatus.SUCCESSFUL);
         expect(screen.getByTestId('grid-card-mOckID1')).toBeVisible();
-        const videoMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
-        expect(videoMenuButton).toBeVisible();
+        const assetMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
+        expect(assetMenuButton).toBeVisible();
 
-        axiosMock.onGet(`${getVideosUrl(courseId)}/mOckID1/usage`).reply(201, { usageLocations: [] });
+        axiosMock.onPut(`${getAssetsUrl(courseId)}mOckID1`).reply(201, { locked: false });
+        axiosMock.onGet(`${getAssetsUrl(courseId)}mOckID1/usage`).reply(201, { usageLocations: [] });
         await waitFor(() => {
-          fireEvent.click(within(videoMenuButton).getByLabelText('file-menu-toggle'));
+          fireEvent.click(within(assetMenuButton).getByLabelText('file-menu-toggle'));
           fireEvent.click(screen.getByText('Info'));
-        });
+          executeThunk(getUsagePaths({
+            courseId,
+            asset: { id: 'mOckID1', displayName: 'mOckID1' },
+            setSelectedRows: jest.fn(),
+          }), store.dispatch);
+          expect(screen.getAllByLabelText('mOckID1')[0]).toBeVisible();
 
+          fireEvent.click(screen.getByLabelText('Checkbox'));
+          executeThunk(updateAssetLock({
+            courseId,
+            assetId: 'mOckID1',
+            locked: false,
+          }), store.dispatch);
+        });
         expect(screen.getByText(messages.usageNotInUseMessage.defaultMessage)).toBeVisible();
 
-        const infoTab = screen.getAllByRole('tab')[0];
-        expect(infoTab).toBeVisible();
-
-        expect(infoTab).toHaveClass('active');
+        const updateStatus = store.getState().assets.updatingStatus;
+        expect(updateStatus).toEqual(RequestStatus.SUCCESSFUL);
       });
 
-      it('should open video info modal and show transcript tab', async () => {
+      it('should unlock asset', async () => {
         renderComponent();
         await mockStore(RequestStatus.SUCCESSFUL);
         expect(screen.getByTestId('grid-card-mOckID1')).toBeVisible();
-        const videoMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
-        expect(videoMenuButton).toBeVisible();
 
-        axiosMock.onGet(`${getVideosUrl(courseId)}/mOckID1/usage`).reply(201, { usageLocations: [] });
+        const assetMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
+        expect(assetMenuButton).toBeVisible();
+
         await waitFor(() => {
-          fireEvent.click(within(videoMenuButton).getByLabelText('file-menu-toggle'));
-          fireEvent.click(screen.getByText('Info'));
+          axiosMock.onPut(`${getAssetsUrl(courseId)}mOckID1`).reply(201, { locked: false });
+          fireEvent.click(within(assetMenuButton).getByLabelText('file-menu-toggle'));
+          fireEvent.click(screen.getByText('Unlock'));
+          executeThunk(updateAssetLock({
+            courseId,
+            assetId: 'mOckID1',
+            locked: false,
+          }), store.dispatch);
         });
+        const updateStatus = store.getState().assets.updatingStatus;
+        expect(updateStatus).toEqual(RequestStatus.SUCCESSFUL);
+      });
 
-        expect(screen.getByText(messages.usageNotInUseMessage.defaultMessage)).toBeVisible();
+      it('should lock asset', async () => {
+        renderComponent();
+        await mockStore(RequestStatus.SUCCESSFUL);
+        expect(screen.getByTestId('grid-card-mOckID3')).toBeVisible();
 
-        const transcriptTab = screen.getAllByRole('tab')[1];
-        await act(async () => {
-          fireEvent.click(transcriptTab);
+        const assetMenuButton = screen.getByTestId('file-menu-dropdown-mOckID3');
+        expect(assetMenuButton).toBeVisible();
+
+        await waitFor(() => {
+          axiosMock.onPut(`${getAssetsUrl(courseId)}mOckID3`).reply(201, { locked: true });
+          fireEvent.click(within(assetMenuButton).getByLabelText('file-menu-toggle'));
+          fireEvent.click(screen.getByText('Lock'));
+          executeThunk(updateAssetLock({
+            courseId,
+            assetId: 'mOckID3',
+            locked: true,
+          }), store.dispatch);
         });
-        expect(transcriptTab).toBeVisible();
-
-        expect(transcriptTab).toHaveClass('active');
+        const updateStatus = store.getState().assets.updatingStatus;
+        expect(updateStatus).toEqual(RequestStatus.SUCCESSFUL);
       });
 
       it('download button should download file', async () => {
@@ -446,18 +429,14 @@ describe('FilesAndUploads', () => {
         await mockStore(RequestStatus.SUCCESSFUL);
         expect(screen.getByTestId('grid-card-mOckID1')).toBeVisible();
 
-        const videoMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
-        expect(videoMenuButton).toBeVisible();
+        const assetMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
+        expect(assetMenuButton).toBeVisible();
 
-        axiosMock.onGet(`${getVideosUrl(courseId)}/mOckID1`).reply(200, { download_link: 'test' });
         await waitFor(() => {
-          fireEvent.click(within(videoMenuButton).getByLabelText('file-menu-toggle'));
+          fireEvent.click(within(assetMenuButton).getByLabelText('file-menu-toggle'));
           fireEvent.click(screen.getByText('Download'));
         });
-
-        const updateStatus = store.getState().videos.updatingStatus;
-
-        expect(updateStatus).toEqual(RequestStatus.SUCCESSFUL);
+        expect(saveAs).toHaveBeenCalled();
       });
 
       it('delete button should delete file', async () => {
@@ -465,21 +444,21 @@ describe('FilesAndUploads', () => {
         await mockStore(RequestStatus.SUCCESSFUL);
         expect(screen.getByTestId('grid-card-mOckID1')).toBeVisible();
 
-        const fileMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
-        expect(fileMenuButton).toBeVisible();
+        const assetMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
+        expect(assetMenuButton).toBeVisible();
 
         await waitFor(() => {
-          axiosMock.onDelete(`${getCoursVideosApiUrl(courseId)}/mOckID1`).reply(204);
-          fireEvent.click(within(fileMenuButton).getByLabelText('file-menu-toggle'));
+          axiosMock.onDelete(`${getAssetsUrl(courseId)}mOckID1`).reply(204);
+          fireEvent.click(within(assetMenuButton).getByLabelText('file-menu-toggle'));
           fireEvent.click(screen.getByTestId('open-delete-confirmation-button'));
           expect(screen.getByText(messages.deleteConfirmationTitle.defaultMessage)).toBeVisible();
 
           fireEvent.click(screen.getByText(messages.deleteFileButtonLabel.defaultMessage));
           expect(screen.queryByText(messages.deleteConfirmationTitle.defaultMessage)).toBeNull();
 
-          executeThunk(deleteVideoFile(courseId, 'mOckID1', 5), store.dispatch);
+          executeThunk(deleteAssetFile(courseId, 'mOckID1', 5), store.dispatch);
         });
-        const deleteStatus = store.getState().videos.deletingStatus;
+        const deleteStatus = store.getState().assets.deletingStatus;
         expect(deleteStatus).toEqual(RequestStatus.SUCCESSFUL);
 
         expect(screen.queryByTestId('grid-card-mOckID1')).toBeNull();
@@ -488,67 +467,31 @@ describe('FilesAndUploads', () => {
 
     describe('api errors', () => {
       it('invalid file size should show error', async () => {
-        const errorMessage = 'File download.png exceeds maximum size of 5 GB.';
+        const errorMessage = 'File download.png exceeds maximum size of 20 MB.';
         renderComponent();
         await mockStore(RequestStatus.SUCCESSFUL);
-        axiosMock.onPost(getCoursVideosApiUrl(courseId)).reply(413, { error: errorMessage });
-        const addFilesButton = screen.getAllByLabelText('file-input')[3];
+        axiosMock.onPost(getAssetsUrl(courseId)).reply(413, { error: errorMessage });
+        const addFilesButton = screen.getByLabelText('file-input');
         await act(async () => {
           userEvent.upload(addFilesButton, file);
-          await executeThunk(addVideoFile(courseId, file), store.dispatch);
+          await executeThunk(addAssetFile(courseId, file, 1), store.dispatch);
         });
-        const addStatus = store.getState().videos.addingStatus;
+        const addStatus = store.getState().assets.addingStatus;
         expect(addStatus).toEqual(RequestStatus.FAILED);
 
         expect(screen.getByText('Error')).toBeVisible();
       });
 
-      it('404 add file should show error', async () => {
+      it('404 upload should show error', async () => {
         renderComponent();
         await mockStore(RequestStatus.SUCCESSFUL);
-        axiosMock.onPost(getCoursVideosApiUrl(courseId)).reply(404);
-        const addFilesButton = screen.getAllByLabelText('file-input')[3];
+        axiosMock.onPost(getAssetsUrl(courseId)).reply(404);
+        const addFilesButton = screen.getByLabelText('file-input');
         await act(async () => {
           userEvent.upload(addFilesButton, file);
-          await executeThunk(addVideoFile(courseId, file), store.dispatch);
+          await executeThunk(addAssetFile(courseId, file, 1), store.dispatch);
         });
-        const addStatus = store.getState().videos.addingStatus;
-        expect(addStatus).toEqual(RequestStatus.FAILED);
-
-        expect(screen.getByText('Error')).toBeVisible();
-      });
-
-      it('404 add thumbnail should show error', async () => {
-        renderComponent();
-        await mockStore(RequestStatus.SUCCESSFUL);
-        axiosMock.onPost(`${getApiBaseUrl()}/video_images/${courseId}/mOckID1`).reply(404);
-        const addThumbnailButton = screen.getByTestId('video-thumbnail-mOckID1');
-        const thumbnail = new File(['test'], 'sOMEUrl.jpg', { type: 'image/jpg' });
-        await act(async () => {
-          fireEvent.click(addThumbnailButton);
-          await executeThunk(addVideoThumbnail({ file: thumbnail, videoId: 'mOckID1', courseId }), store.dispatch);
-        });
-        const updateStatus = store.getState().videos.updatingStatus;
-        expect(updateStatus).toEqual(RequestStatus.FAILED);
-
-        expect(screen.getByText('Error')).toBeVisible();
-      });
-
-      it('404 upload file to server should show error', async () => {
-        renderComponent();
-        await mockStore(RequestStatus.SUCCESSFUL);
-        const mockResponseData = { status: '404', ok: false, blob: () => 'Data' };
-        const mockFetchResponse = Promise.reject(mockResponseData);
-        global.fetch = jest.fn().mockImplementation(() => mockFetchResponse);
-
-        axiosMock.onPost(getCoursVideosApiUrl(courseId)).reply(204, generateNewVideoApiResponse());
-        axiosMock.onGet(getCoursVideosApiUrl(courseId)).reply(200, generateAddVideoApiResponse());
-        const addFilesButton = screen.getAllByLabelText('file-input')[3];
-        await act(async () => {
-          userEvent.upload(addFilesButton, file);
-          await executeThunk(addVideoFile(courseId, file), store.dispatch);
-        });
-        const addStatus = store.getState().videos.addingStatus;
+        const addStatus = store.getState().assets.addingStatus;
         expect(addStatus).toEqual(RequestStatus.FAILED);
 
         expect(screen.getByText('Error')).toBeVisible();
@@ -559,21 +502,21 @@ describe('FilesAndUploads', () => {
         await mockStore(RequestStatus.SUCCESSFUL);
         expect(screen.getByTestId('grid-card-mOckID1')).toBeVisible();
 
-        const videoMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
-        expect(videoMenuButton).toBeVisible();
+        const assetMenuButton = screen.getByTestId('file-menu-dropdown-mOckID1');
+        expect(assetMenuButton).toBeVisible();
 
         await waitFor(() => {
-          axiosMock.onDelete(`${getCoursVideosApiUrl(courseId)}/mOckID1`).reply(404);
-          fireEvent.click(within(videoMenuButton).getByLabelText('file-menu-toggle'));
+          axiosMock.onDelete(`${getAssetsUrl(courseId)}mOckID1`).reply(404);
+          fireEvent.click(within(assetMenuButton).getByLabelText('file-menu-toggle'));
           fireEvent.click(screen.getByTestId('open-delete-confirmation-button'));
           expect(screen.getByText(messages.deleteConfirmationTitle.defaultMessage)).toBeVisible();
 
           fireEvent.click(screen.getByText(messages.deleteFileButtonLabel.defaultMessage));
           expect(screen.queryByText(messages.deleteConfirmationTitle.defaultMessage)).toBeNull();
 
-          executeThunk(deleteVideoFile(courseId, 'mOckID1', 5), store.dispatch);
+          executeThunk(deleteAssetFile(courseId, 'mOckID1', 5), store.dispatch);
         });
-        const deleteStatus = store.getState().videos.deletingStatus;
+        const deleteStatus = store.getState().assets.deletingStatus;
         expect(deleteStatus).toEqual(RequestStatus.FAILED);
 
         expect(screen.getByTestId('grid-card-mOckID1')).toBeVisible();
@@ -586,43 +529,71 @@ describe('FilesAndUploads', () => {
         await mockStore(RequestStatus.SUCCESSFUL);
         expect(screen.getByTestId('grid-card-mOckID3')).toBeVisible();
 
-        const videoMenuButton = screen.getByTestId('file-menu-dropdown-mOckID3');
-        expect(videoMenuButton).toBeVisible();
+        const assetMenuButton = screen.getByTestId('file-menu-dropdown-mOckID3');
+        expect(assetMenuButton).toBeVisible();
 
-        axiosMock.onGet(`${getVideosUrl(courseId)}/mOckID3/usage`).reply(404);
+        axiosMock.onGet(`${getAssetsUrl(courseId)}mOckID3/usage`).reply(404);
         await waitFor(() => {
-          fireEvent.click(within(videoMenuButton).getByLabelText('file-menu-toggle'));
+          fireEvent.click(within(assetMenuButton).getByLabelText('file-menu-toggle'));
           fireEvent.click(screen.getByText('Info'));
           executeThunk(getUsagePaths({
             courseId,
-            video: { id: 'mOckID3', displayName: 'mOckID3' },
+            asset: { id: 'mOckID3', displayName: 'mOckID3' },
+            setSelectedRows: jest.fn(),
           }), store.dispatch);
         });
-        const { usageStatus } = store.getState().videos;
+        const { usageStatus } = store.getState().assets;
         expect(usageStatus).toEqual(RequestStatus.FAILED);
       });
 
-      it('multiple video files fetch failure should show error', async () => {
+      it('404 lock update should show error', async () => {
+        renderComponent();
+        await mockStore(RequestStatus.SUCCESSFUL);
+        expect(screen.getByTestId('grid-card-mOckID3')).toBeVisible();
+
+        const assetMenuButton = screen.getByTestId('file-menu-dropdown-mOckID3');
+        expect(assetMenuButton).toBeVisible();
+
+        await waitFor(() => {
+          axiosMock.onPut(`${getAssetsUrl(courseId)}mOckID3`).reply(404);
+          fireEvent.click(within(assetMenuButton).getByLabelText('file-menu-toggle'));
+          fireEvent.click(screen.getByText('Lock'));
+          executeThunk(updateAssetLock({
+            courseId,
+            assetId: 'mOckID3',
+            locked: true,
+          }), store.dispatch);
+        });
+        const updateStatus = store.getState().assets.updatingStatus;
+        expect(updateStatus).toEqual(RequestStatus.FAILED);
+
+        expect(screen.getByText('Error')).toBeVisible();
+      });
+
+      it('multiple asset file fetch failure should show error', async () => {
         renderComponent();
         await mockStore(RequestStatus.SUCCESSFUL);
         const selectCardButtons = screen.getAllByTestId('datatable-select-column-checkbox-cell');
         fireEvent.click(selectCardButtons[0]);
-        fireEvent.click(selectCardButtons[2]);
+        fireEvent.click(selectCardButtons[1]);
         const actionsButton = screen.getByText(messages.actionsButtonLabel.defaultMessage);
         expect(actionsButton).toBeVisible();
 
         await waitFor(() => {
           fireEvent.click(actionsButton);
         });
+        const mockResponseData = { ok: false };
+        const mockFetchResponse = Promise.resolve(mockResponseData);
         const downloadButton = screen.getByText(messages.downloadTitle.defaultMessage).closest('a');
         expect(downloadButton).not.toHaveClass('disabled');
 
+        global.fetch = jest.fn().mockImplementation(() => mockFetchResponse);
         await waitFor(() => {
           fireEvent.click(downloadButton);
-          executeThunk(fetchVideoDownload([{ original: { displayName: 'mOckID1', id: '2' } }]), store.dispatch);
+          expect(fetch).toHaveBeenCalledTimes(2);
         });
 
-        const updateStatus = store.getState().videos.updatingStatus;
+        const updateStatus = store.getState().assets.updatingStatus;
         expect(updateStatus).toEqual(RequestStatus.FAILED);
 
         expect(screen.getByText('Error')).toBeVisible();
