+++ src/files-and-videos/generic/EditFileErrors.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { injectIntl, intlShape } from '@edx/frontend-platform/i18n';
 import { ErrorAlert } from '@edx/frontend-lib-content-components';
-import { RequestStatus } from '../data/constants';
+import { RequestStatus } from '../../data/constants';
 import messages from './messages';
 
 const EditFileErrors = ({
