+++ src/files-and-videos/generic/index.js
@@ -1,0 +1,24 @@
+import {
+  TableActions,
+  GalleryCard,
+  AccessColumn,
+  ActiveColumn,
+  MoreInfoColumn,
+  StatusColumn,
+  ThumbnailColumn,
+} from './table-components';
+import FileInput, { useFileInput } from './FileInput';
+
+export {
+  TableActions,
+  GalleryCard,
+  AccessColumn,
+  ActiveColumn,
+  MoreInfoColumn,
+  StatusColumn,
+  ThumbnailColumn,
+  FileInput,
+  useFileInput,
+};
+export { default as FileTable } from './FileTable';
+export { default as EditFileErrors } from './EditFileErrors';
