+++ src/files-and-videos/generic/table-components/GalleryCard.scss
