+++ src/files-and-videos/generic/table-components/index.js
@@ -1,0 +1,19 @@
+import GalleryCard from './GalleryCard';
+import TableActions from './TableActions';
+import {
+  AccessColumn,
+  ActiveColumn,
+  MoreInfoColumn,
+  StatusColumn,
+  ThumbnailColumn,
+} from './table-custom-columns';
+
+export {
+  TableActions,
+  GalleryCard,
+  AccessColumn,
+  ActiveColumn,
+  MoreInfoColumn,
+  StatusColumn,
+  ThumbnailColumn,
+};
