+++ src/files-and-videos/index.scss
@@ -1,3 +1,3 @@
 @import "files-and-videos/videos-page/transcript-settings/TranscriptSettings";
 @import "files-and-videos/videos-page/VideoThumbnail";
-@import "files-and-videos/table-components/GalleryCard"
+@import "files-and-videos/generic/table-components/GalleryCard"
