--- src/files-and-videos/table-components/index.js
@@ -1,7 +1,0 @@
-import GalleryCard from './GalleryCard';
-import TableActions from './TableActions';
-
-export {
-  TableActions,
-  GalleryCard,
-};
