+++ src/files-and-videos/videos-page/info-sidebar/index.js
@@ -1,0 +1,3 @@
+import VideoInfoModalSidebar from './VideoInfoModalSidebar';
+
+export default VideoInfoModalSidebar;
